+++ src/basic/hooks/useCart.ts
@@ -23,18 +23,117 @@
 // - calculateTotal: 총액 계산 함수
 // - getRemainingStock: 재고 확인 함수
 // - clearCart: 장바구니 비우기 함수
-import { CartItem, Product } from '../types';
+import { useCallback, useState, useEffect } from 'react';
 
-const useCart = () => {
-  const getRemainingStock = (product: Product, cart: CartItem[]): number => {
+import { updateCartItemQuantity } from '../models/cart';
+import { CartItem, Product, ProductWithUI } from '../types';
+
+const useCart = (
+  products: Product[],
+  addNotification: (message: string, type?: 'error' | 'success' | 'warning') => void,
+) => {
+  const [cart, setCart] = useState<CartItem[]>(() => {
+    const saved = localStorage.getItem('cart');
+    if (saved) {
+      try {
+        return JSON.parse(saved);
+      } catch {
+        return [];
+      }
+    }
+    return [];
+  });
+  const [totalItemCount, setTotalItemCount] = useState(0);
+
+  useEffect(() => {
+    const count = cart.reduce((sum, item) => sum + item.quantity, 0);
+    setTotalItemCount(count);
+  }, [cart]);
+
+  useEffect(() => {
+    if (cart.length > 0) {
+      localStorage.setItem('cart', JSON.stringify(cart));
+    } else {
+      localStorage.removeItem('cart');
+    }
+  }, [cart]);
+
+  const getRemainingStock = (product: Product): number => {
     const cartItem = cart.find((item) => item.product.id === product.id);
     const remaining = product.stock - (cartItem?.quantity || 0);
 
     return remaining;
   };
 
+  const addToCart = useCallback(
+    (product: ProductWithUI) => {
+      const remainingStock = getRemainingStock(product);
+      if (remainingStock <= 0) {
+        addNotification('재고가 부족합니다!', 'error');
+        return;
+      }
+
+      setCart((prevCart) => {
+        const existingItem = prevCart.find((item) => item.product.id === product.id);
+
+        if (existingItem) {
+          const newQuantity = existingItem.quantity + 1;
+
+          if (newQuantity > product.stock) {
+            addNotification(`재고는 ${product.stock}개까지만 있습니다.`, 'error');
+            return prevCart;
+          }
+
+          return prevCart.map((item) =>
+            item.product.id === product.id ? { ...item, quantity: newQuantity } : item,
+          );
+        }
+
+        return [...prevCart, { product, quantity: 1 }];
+      });
+
+      addNotification('장바구니에 담았습니다', 'success');
+    },
+    [cart, addNotification, getRemainingStock],
+  );
+
+  const removeFromCart = useCallback((productId: string) => {
+    setCart((prevCart) => prevCart.filter((item) => item.product.id !== productId));
+  }, []);
+
+  const updateQuantity = useCallback(
+    (productId: string, newQuantity: number) => {
+      if (newQuantity <= 0) {
+        setCart((prevCart) => updateCartItemQuantity(prevCart, productId, 0));
+        return;
+      }
+
+      const product = products.find((p) => p.id === productId);
+      if (!product) return;
+
+      const maxStock = product.stock;
+      if (newQuantity > maxStock) {
+        addNotification(`재고는 ${maxStock}개까지만 있습니다.`, 'error');
+        return;
+      }
+
+      setCart((prevCart) => updateCartItemQuantity(prevCart, productId, newQuantity));
+    },
+    [setCart, products, addNotification],
+  );
+
+  const clearCart = useCallback(() => {
+    setCart([]);
+  }, []);
+
   return {
+    cart,
     getRemainingStock,
+    updateQuantity,
+    addToCart,
+    removeFromCart,
+    totalItemCount,
+    clearCart,
   };
 };
