+++ src/basic/hooks/useNotification.ts
@@ -1,0 +1,31 @@
+import { useCallback, useState } from 'react';
+
+import { Notification } from '../types';
+
+const useNotification = () => {
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+
+  const addNotification = useCallback(
+    (message: string, type: 'error' | 'success' | 'warning' = 'success') => {
+      const id = Date.now().toString();
+      setNotifications((prev) => [...prev, { id, message, type }]);
+
+      setTimeout(() => {
+        setNotifications((prev) => prev.filter((n) => n.id !== id));
+      }, 3000);
+    },
+    [],
+  );
+
+  const removeNotification = useCallback((id: string) => {
+    setNotifications((prev) => prev.filter((n) => n.id !== id));
+  }, []);
+
+  return {
+    notifications,
+    addNotification,
+    removeNotification,
+  };
+};
+
+export { useNotification };
